Issue #35: Removed wrong colors from svg arts. Moved all used svgs to `src/assets/arts`.
